# gateway/internal/gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 多平台音乐网关 API
  version: 0.5.3
  description: |
    **多平台音乐网关 API** 是一组用于聚合 QQ 音乐 / 网易云音乐 / 酷狗音乐等多家音乐源的统一 RESTful 接口。
    网关在内部完成账号池调度、限流配额、计费扣减、Extra Usage 超额计费、余额扣款等控制逻辑，
    上游应用只需要通过统一的 API 接入即可。

    ## 协议总览
    - **基础路径**：所有业务接口位于 `/v1/...` 命名空间下。
    - **认证方式**：
      - **Cookie 会话**（前端控制台用）：登录后服务器下发 `sid` 会话 Cookie。
      - **API Key**（程序化调用用）：在 `Authorization: Bearer <api_key>` 请求头中携带。
    - **响应封装**：网关接口绝大多数返回 `{ code, message, data }` 包装结构（`code = 0` 表示成功），
      详情见下方 `Response` schema 定义。
    - **CSRF 保护**：使用 Cookie 会话时，所有非 GET/HEAD/OPTIONS 请求都需要在请求头中携带
      `X-CSRF-Token`（值与 Cookie `csrf_token` 一致）。API Key 模式不受 CSRF 限制。
    - **限流模型**：
      - **QPS**：按用户套餐限制每秒请求数。
      - **每日 / 每月配额**：免费版默认 100 次/日；超出后若开启 Extra Usage 则会按千次价从余额中扣费。

    ## 错误码
    全局错误码定义在 `ErrorCode` 字段：
    | code | 含义 |
    | ---- | ---- |
    | 0 | 成功 |
    | 400 | 请求参数错误 |
    | 401 | 未认证 / 凭据失效 |
    | 403 | 权限不足 |
    | 404 | 资源不存在 |
    | 409 | 资源冲突（重复创建、版本冲突等） |
    | 429 | 限流（QPS 或日/月配额耗尽） |
    | 500 | 服务器内部错误 |
  contact:
    name: 多平台音乐网关
    url: https://github.com/luoye663/qqmusicapi
servers:
  - url: /
    description: 当前实例
  - url: https://api.example.com
    description: 生产环境（占位，按部署实际域名替换）
...
security:
  - cookieAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/{provider}/songs/{id}:
    parameters:
      - $ref: "#/components/parameters/Provider"
      - { name: id, in: path, required: true, schema: { type: string }, description: 歌曲 ID }
    get:
      tags: [音乐]
      summary: 获取歌曲详情
      responses:
        "200":
          description: 歌曲详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SongEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/{provider}/songs/{id}/url:
    parameters:
      - $ref: "#/components/parameters/Provider"
      - { name: id, in: path, required: true, schema: { type: string } }
      - { name: quality, in: query, schema: { $ref: "#/components/schemas/AudioQuality" }, description: "音质，默认 MP3_320" }
    get:
      tags: [音乐]
      summary: 获取歌曲直链（带防盗链）
      responses:
        "200":
          description: 直链
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SongURLEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/{provider}/songs/{id}/lyric:
    parameters:
      - $ref: "#/components/parameters/Provider"
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [音乐]
      summary: 获取歌词
      responses:
        "200":
          description: 歌词
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LyricEnvelope" }
  /v1/{provider}/search/songs:
    parameters:
      - $ref: "#/components/parameters/Provider"
      - { name: q, in: query, required: true, schema: { type: string }, description: 搜索关键词 }
      - { name: page, in: query, schema: { type: integer, default: 1 } }
      - { name: page_size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
    get:
      tags: [音乐]
      summary: 搜索歌曲
      responses:
        "200":
          description: 搜索结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SongSearchEnvelope" }
  /v1/{provider}/albums/{id}:
    parameters:
      - $ref: "#/components/parameters/Provider"
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [音乐]
      summary: 获取专辑
      responses:
        "200":
          description: 专辑
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlbumEnvelope" }
  /v1/{provider}/artists/{id}:
    parameters:
      - $ref: "#/components/parameters/Provider"
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [音乐]
      summary: 获取歌手
      responses:
        "200":
          description: 歌手
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtistEnvelope" }
  /v1/{provider}/playlists/{id}:
    parameters:
      - $ref: "#/components/parameters/Provider"
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [音乐]
      summary: 获取歌单
      responses:
        "200":
          description: 歌单
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlaylistEnvelope" }

  # ========== 网易云原生登录 ==========
  /v1/netease/auth/login/cellphone:
    post:
      tags: [网易云登录]
      summary: 手机号登录
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NeteaseCellphoneLogin" }
      responses:
        "200": { $ref: "#/components/responses/RawJSON" }
  /v1/netease/auth/login/email:
    post:
      tags: [网易云登录]
      summary: 邮箱登录
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NeteaseEmailLogin" }
      responses:
        "200": { $ref: "#/components/responses/RawJSON" }
...
  /v1/netease/auth/login/qr/create:
    post:
      tags: [网易云登录]
      summary: 创建二维码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key: { type: string }
                qrimg: { type: boolean }
      responses: { "200": { $ref: "#/components/responses/RawJSON" } }
  /v1/netease/auth/login/qr/check:
    post:
      tags: [网易云登录]
      summary: 检测二维码扫码状态
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key: { type: string }
      responses: { "200": { $ref: "#/components/responses/RawJSON" } }
...
  /v1/netease/auth/captcha/sent:
    post:
      tags: [网易云登录]
      summary: 发送短信验证码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string }
                countryCode: { type: string }
      responses: { "200": { $ref: "#/components/responses/RawJSON" } }
  /v1/netease/auth/captcha/verify:
    post:
      tags: [网易云登录]
      summary: 校验短信验证码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, captcha]
              properties:
                phone: { type: string }
                captcha: { type: string }
                countryCode: { type: string }
      responses: { "200": { $ref: "#/components/responses/RawJSON" } }

  # ========== 用户管理（管理员） ==========
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: 通过 `/v1/auth/login` 颁发的会话 Cookie，浏览器自动携带；同源调用无需手动配置。
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: "在 `Authorization: Bearer <api_key>` 中携带。"
  parameters:
    Provider:
      name: provider
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Provider" }
      description: 上游音乐源标识。
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    RateLimited:
      description: 触发限流（QPS / 每日 / 每月配额耗尽）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    RawJSON:
      description: 直接透传上游返回，未做封装
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  schemas:
...
    Response:
      type: object
      required: [code, message]
      properties:
        code: { type: integer, example: 0, description: "0=成功；非 0 见错误码表" }
        message: { type: string, example: "ok" }
        data: { description: "实际业务数据，类型见各接口" }
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          example: { code: 400, message: "invalid body", data: null }
...
    Provider:
      type: string
      enum: [qqmusic, netease, kugou]
      description: 音乐源
    AudioQuality:
      type: string
      enum: [MP3_128, MP3_320, FLAC, ATMOS, ATMOS2]
      description: 音质标签

    # ---- 用户 / 认证 ----
...
    Song:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        subtitle: { type: string }
        duration: { type: integer, format: int64, description: "毫秒" }
        artists:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
        album:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            picUrl: { type: string }
        publishTime: { type: string, format: date-time, nullable: true }
        provider: { $ref: "#/components/schemas/Provider" }
    SongEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Song" }
    SongURL:
      type: object
      properties:
        id: { type: string }
        url: { type: string, format: uri }
        format: { type: string, example: "mp3" }
        bitrate: { type: integer }
        size: { type: integer, format: int64 }
        expiresAt: { type: string, format: date-time, nullable: true }
    SongURLEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/SongURL" }
    Lyric:
      type: object
      properties:
        lrc: { type: string, description: "LRC 格式歌词" }
        trans: { type: string, description: "翻译歌词" }
    LyricEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Lyric" }
    SongSearchEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/Song" } }
                total: { type: integer }
                page: { type: integer }
                pageSize: { type: integer }
                hasMore: { type: boolean }
    Album:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        picUrl: { type: string }
        publishTime: { type: string, format: date-time, nullable: true }
        artists: { type: array, items: { type: object, properties: { id: { type: string }, name: { type: string } } } }
        songs: { type: array, items: { $ref: "#/components/schemas/Song" } }
    AlbumEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Album" }
    Artist:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        picUrl: { type: string }
        description: { type: string }
        songs: { type: array, items: { $ref: "#/components/schemas/Song" } }
    ArtistEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Artist" }
    Playlist:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        creator: { type: string }
        picUrl: { type: string }
        songs: { type: array, items: { $ref: "#/components/schemas/Song" } }
    PlaylistEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Playlist" }

    # ---- 网易云原生登录 ----
    NeteaseCellphoneLogin:
      type: object
      required: [phone]
      properties:
        phone: { type: string }
        password: { type: string }
        md5Password: { type: string }
        captcha: { type: string }
        countryCode: { type: string }
    NeteaseEmailLogin:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        password: { type: string }
        md5Password: { type: string }

    # ---- 用户管理（管理员） ----
...